refactor#11 bottom sheet package
