+++ feature/maps/src/main/java/com/teamsolply/solply/maps/bottomsheet/place/PlaceBottomSheet.kt
@@ -1,0 +1,6 @@
+package com.teamsolply.solply.maps.bottomsheet.place
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun PlaceBottomSheet(){}
